Fix: Using System.DoubleNumerics for Transforms (#2800) (#2815)
* Added System.DoubleNumerics to the core project
Converted Transform.cs to use System.DoubleNumerics
Added a new TransformTests.cs file to test the new transform class
Changed Converters to use System.DoubleNumerics

* Adds handling of new System.DoubleNumerics matrix4x4 to serializer and deserializer

* Update ConverterNavisworks.Geometry.cs

* updates transform methods in rhino and revit

* test(core): Tests matrix serialisation

* Added log for backwards compatibility usage

---------

+++ Objects/Converters/ConverterRevit/ConverterRevitShared/Partial Classes/ConvertFamilyInstance.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Numerics;
+using System.DoubleNumerics;
 
 using Autodesk.Revit.DB;
 using Autodesk.Revit.DB.Structure;
@@ -456,7 +456,7 @@ namespace Objects.Converter.Revit
       );
 
       // get the scale: TODO: do revit transforms ever have scaling?
-      var scale = (float)transform.Scale;
+      var scale = transform.Scale;
 
       return new Other.Transform(vX, vY, vZ, t) { units = ModelUnits };
     }
